Agentic AI With Langgraph And MCP/pyproject.toml
[project]
name = "agentic-ai-with-langgraph-and-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-google-community[calendar]>=2.0.7",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.4.8",
    "langsmith>=0.4.1",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
]
